% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC.tsdistribution.estimate}
\alias{AIC.tsdistribution.estimate}
\alias{AIC}
\title{Akaike's An Information Criterion}
\usage{
\method{AIC}{tsdistribution.estimate}(object, ..., k = 2)
}
\arguments{
\item{object}{an object of class \dQuote{tsdistribution.estimate}.}

\item{...}{not currently used.}

\item{k}{the penalty per parameter to be used; the default k = 2 is the
classical AIC.}
}
\value{
The AIC value (scalar).
}
\description{
Extract the AIC from an estimated model.
}
