% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC.tsdistribution.estimate}
\alias{BIC.tsdistribution.estimate}
\alias{BIC}
\title{Bayesian Information Criterion}
\usage{
\method{BIC}{tsdistribution.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsdistribution.estimate}.}

\item{...}{not currently used.}
}
\value{
The BIC value (scalar).
}
\description{
Extract the BIC from an estimated model.
}
