% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.tsdistribution.estimate}
\alias{coef.tsdistribution.estimate}
\alias{coef}
\title{Extract Model Coefficients}
\usage{
\method{coef}{tsdistribution.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class tsdistribution.estimate.}

\item{...}{other arguments.}
}
\value{
The function computes and returns a list of summary statistics of
the fitted model given in object.
}
\description{
Extract Model Coefficients
}
