% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution_bounds}
\alias{distribution_bounds}
\title{Distribution Bounds}
\usage{
distribution_bounds(distribution = "norm")
}
\arguments{
\item{distribution}{A valid distribution}
}
\value{
A data.table of the parameters and their default bounds.
}
\description{
Distribution Bounds
}
\details{
Returns the upper a lower bounds for the parameters of a distribution.
}
