% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specification.R
\name{distribution_modelspec}
\alias{distribution_modelspec}
\title{Specification of distribution model}
\usage{
distribution_modelspec(y, distribution = "norm", ...)
}
\arguments{
\item{y}{a numeric vector}

\item{distribution}{the type of distribution. Valid choices are norm (Normal),
snorm (Skew Normal), std (Student), sstd (Skew Student), ged (Generalized Error),
sged (Skew Generalized Error), nig (Normal Inverse Gaussian), gh (Generalized Hyperbolic),
ghst (Generalized Hyperbolic Skew Student) and jsu (Johnson's SU).}

\item{...}{not currently used}
}
\value{
An object of class tsdistribution.spec
}
\description{
Specification of distribution model
}
