% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{dskewness}
\alias{dskewness}
\alias{dkurtosis}
\title{Distribution skewness and kurtosis}
\usage{
dskewness(distribution = "norm", skew = 1, shape = 5, lambda = -0.5)

dkurtosis(distribution = "norm", skew = 1, shape = 5, lambda = -0.5)
}
\arguments{
\item{distribution}{a valid distribution.}

\item{skew}{skew parameter.}

\item{shape}{shape parameter.}

\item{lambda}{additional shape parameter for the Generalized Hyperbolic
distribution.}
}
\value{
A numeric value.
}
\description{
Calculates the skewness and excess kurtosis of the distribution given
a set of parameters.
}
