% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandwich.R
\name{estfun.tsdistribution.estimate}
\alias{estfun.tsdistribution.estimate}
\alias{estfun}
\title{Score Method}
\usage{
\method{estfun}{tsdistribution.estimate}(x, ...)
}
\arguments{
\item{x}{an object of class \dQuote{tsdistribution.estimate}.}

\item{...}{not currently used.}
}
\value{
The score matrix
}
\description{
Score Method
}
\details{
The function returns the scores of likelihood at the optimal solution.
}
\author{
Alexios Galanos
}
