% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr.R
\name{dged}
\alias{dged}
\alias{pged}
\alias{qged}
\alias{rged}
\title{Generalized Error Distribution}
\usage{
dged(x, mu = 0, sigma = 1, shape = 2, log = FALSE)

pged(q, mu = 0, sigma = 1, shape = 2, lower_tail = TRUE, log = TRUE)

qged(p, mu = 0, sigma = 1, shape = 2, lower_tail = TRUE, log = FALSE)

rged(n, mu = 0, sigma = 1, shape = 2)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mean.}

\item{sigma}{standard deviation.}

\item{shape}{shape parameter.}

\item{log}{(logical) if TRUE, probabilities p are given as log(p).}

\item{lower_tail}{if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{Number of observations.}
}
\value{
d gives the density, p gives the distribution function, q gives the quantile function 
and r generates random deviates. Output depends on x or q length, or n for the random number 
generator
}
\description{
Density, distribution, quantile function and random number 
generation for the location scale invariant parameterization of the 
generalized error distribution.
}
