% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr.R
\name{dsghst}
\alias{dsghst}
\alias{rsghst}
\alias{psghst}
\alias{qsghst}
\title{Skewed Generalized Hyperbolic Student Distribution}
\usage{
dsghst(x, mu = 0, sigma = 1, skew = 1, shape = 8, log = FALSE)

rsghst(n, mu = 0, sigma = 1, skew = 1, shape = 8)

psghst(
  q,
  mu = 0,
  sigma = 1,
  skew = 1,
  shape = 8,
  lower_tail = TRUE,
  log = FALSE
)

qsghst(
  p,
  mu = 0,
  sigma = 1,
  skew = 1,
  shape = 8,
  lower_tail = TRUE,
  log = FALSE
)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mean.}

\item{sigma}{standard deviation.}

\item{skew}{skew parameter.}

\item{shape}{shape parameter.}

\item{log}{(logical) if TRUE, probabilities p are given as log(p).}

\item{n}{Number of observations.}

\item{lower_tail}{if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{vector of probabilities.}
}
\value{
d gives the density, p gives the distribution function, q gives the quantile function 
and r generates random deviates. Output depends on x or q length, or n for the random number 
generator
}
\description{
Density, distribution, quantile function and random number 
generation for the location scale invariant parameterization of the 
generalized hyperbolic student distribution.
}
