% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr.R
\name{nigtransform}
\alias{nigtransform}
\alias{ghyptransform}
\title{Parameter Transformation}
\usage{
nigtransform(mu = 0, sigma = 1, skew = 0, shape = 3)

ghyptransform(mu = 0, sigma = 1, skew = 0, shape = 3, lambda = -0.5)
}
\arguments{
\item{mu}{mean.}

\item{sigma}{standard deviation.}

\item{skew}{skew parameter.}

\item{shape}{shape parameter.}

\item{lambda}{additional shape parameter for the Generalized Hyperbolic
distribution.}
}
\value{
The (alpha, beta, delta, mu) representation.
}
\description{
Transforms parameters from standardized representation to distribution
specific representation for the nig and gh distributions.
}
