% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{logLik.tsdistribution.estimate}
\alias{logLik.tsdistribution.estimate}
\alias{logLik}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{tsdistribution.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class tsdistribution.estimate.}

\item{...}{other arguments.}
}
\value{
An object of class logLik. This is a number with at least 
one attribute, \dQuote{df} (degrees of freedom), giving the number of 
(estimated) parameters in the model.
}
\description{
Extract Log-Likelihood
}
