% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.tsdistribution}
\alias{print.summary.tsdistribution}
\title{Model Estimation Summary Print method}
\usage{
\method{print}{summary.tsdistribution}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  table.caption = paste0(toupper(x$distribution), " Model Summary\\n"),
  ...
)
}
\arguments{
\item{x}{an object of class \dQuote{summary.tsdistribution}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{signif.stars}{logical. If TRUE, ‘significance stars’ are printed for each coefficient.}

\item{table.caption}{an optional string for the table caption.}

\item{...}{not currently used.}
}
\value{
Console output of the object summary.
}
\description{
Print method for class \dQuote{summary.tsdistribution}
}
