% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{print.summary.tsdistribution.profile}
\alias{print.summary.tsdistribution.profile}
\title{Profile Summary Print method}
\usage{
\method{print}{summary.tsdistribution.profile}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an object of class \dQuote{summary.tsdistribution.profile}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{...}{not currently used.}
}
\value{
Invisibly returns the original summary object and prints out to the console.
}
\description{
Print method for class \dQuote{summary.tsdistribution.profile}
}
