% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdqr.R
\name{dsgh}
\alias{dsgh}
\alias{psgh}
\alias{qsgh}
\alias{rsgh}
\title{(Standardized) Generalized Hyperbolic Distribution}
\usage{
dsgh(x, mu = 0, sigma = 1, skew = 0, shape = 1, lambda = 1, log = FALSE)

psgh(
  q,
  mu = 0,
  sigma = 1,
  skew = 0,
  shape = 1,
  lambda = 1,
  lower_tail = TRUE,
  log = FALSE
)

qsgh(
  p,
  mu = 0,
  sigma = 1,
  skew = 0,
  shape = 1,
  lambda = 1,
  lower_tail = TRUE,
  log = FALSE
)

rsgh(n, mu = 0, sigma = 1, skew = 0, shape = 1, lambda = 1)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{mu}{mean.}

\item{sigma}{standard deviation.}

\item{skew}{skew parameter.}

\item{shape}{shape parameter.}

\item{lambda}{additional shape parameter determining subfamilies of this 
distributions.}

\item{log}{(logical) if TRUE, probabilities p are given as log(p).}

\item{lower_tail}{if TRUE (default), probabilities are P[X ≤ x] otherwise, P[X > x].}

\item{p}{vector of probabilities.}

\item{n}{number of observations.}
}
\value{
d gives the density, p gives the distribution function, q gives the quantile function 
and r generates random deviates. Output depends on x or q length, or n for the random number 
generator
}
\description{
Density, distribution, quantile function and random number 
generation for the location scale invariant parameterization of the 
standardized generalized hyperbolic distribution.
}
