% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{summary.tsdistribution.profile}
\alias{summary.tsdistribution.profile}
\title{Distribution Profile Summary}
\usage{
\method{summary}{tsdistribution.profile}(object, digits = 4, measure = "RMSE", ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsdistribution.profile}.}

\item{digits}{integer, used for number formatting. Optionally, to avoid
scientific notation, set \sQuote{options(scipen=999)}.}

\item{measure}{either one of the 3 included measure in the summary slot of
the returned object \dQuote{RMSE}, \dQuote{MAE} or \dQuote{MAPE}, else any
other user calculated measure which has been generated in the summary table
post processing.}

\item{...}{not currently used.}
}
\value{
A list with summary information of class \dQuote{summary.tsdistribution.profile},
including a table with each actual parameter against the measure chosen across each
size in the profile.
}
\description{
Summary method for class \dQuote{tsdistribution.profile}
}
