% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd.R
\name{summary.tsdistribution.spdestimate}
\alias{summary.tsdistribution.spdestimate}
\title{Summary of estimated SPD distribution}
\usage{
\method{summary}{tsdistribution.spdestimate}(object, ...)
}
\arguments{
\item{object}{an object of class \dQuote{tsdistribution.spdestimate}.}

\item{...}{additional parameters passed to the summary method.}
}
\value{
A list of summary statistics of the fitted model given in object.
}
\description{
Summary of estimated SPD distribution
}
\details{
The standard errors assume a blog diagonal covariance structure between
the upper and lower Generalized Pareto Tails.
}
