% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdistributions-package.R
\docType{package}
\name{tsdistributions-package}
\alias{tsdistributions}
\alias{tsdistributions-package}
\title{tsdistributions: Univariate Standardized Distributions}
\description{
A numbers of reparameterized distributions in mean, variance space with full functionality (p,q,r,d) and estimation using automatic differentiation.
}
\keyword{internal}
