% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tsmoments.tsdistribution.estimate}
\alias{tsmoments.tsdistribution.estimate}
\alias{tsmoments}
\title{Extract the moments of an estimated distribution}
\usage{
\method{tsmoments}{tsdistribution.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class tsdistribution.estimate.}

\item{...}{other arguments.}
}
\value{
A vector of the first four moments of the distribution based on the 
estimated parameters. The kurtosis represents the value in excess of 3.
}
\description{
Extract the moments of an estimated distribution
}
