% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.tsdistribution.estimate}
\alias{coef.tsdistribution.estimate}
\alias{coef}
\title{Extract Model Coefficients}
\usage{
\method{coef}{tsdistribution.estimate}(object, ...)
}
\arguments{
\item{object}{an object of class tsdistribution.estimate.}

\item{...}{other arguments.}
}
\value{
A vector of the estimated model coefficients.
}
\description{
Extract Model Coefficients
}
