% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdistributions-package.R
\docType{package}
\name{tsdistributions-package}
\alias{tsdistributions}
\alias{tsdistributions-package}
\title{tsdistributions: Location Scale Standardized Distributions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Location-Scale based distributions parameterized in terms of mean, standard deviation, skew and shape parameters and estimation using automatic differentiation. Distributions include the Normal, Student and GED as well as their skewed variants ('Fernandez and Steel'), the 'Johnson SU', and the Generalized Hyperbolic. Also included is the semi-parametric piece wise distribution ('spd') with Pareto tails and kernel interior.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nopredict.com/packages/tsdistributions}
  \item \url{https://github.com/tsmodels/tsdistributions}
}

}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} (\href{https://orcid.org/0009-0000-9308-0457}{ORCID}) [copyright holder]

}
\keyword{internal}
