% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsdistributions-package.R
\docType{package}
\name{tsdistributions-package}
\alias{tsdistributions}
\alias{tsdistributions-package}
\title{tsdistributions: Location Scale Standardized Distributions}
\description{
Location-Scale based distributions parameterized in terms of mean, standard deviation, skew and shape parameters and estimation using automatic differentiation. Distributions include the Normal, Student and GED as well as their skewed variants ('Fernandez and Steel'), the 'Johnson SU', and the Generalized Hyperbolic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.nopredict.com/packages/tsdistributions}
  \item \url{https://github.com/tsmodels/tsdistributions}
}

}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} [copyright holder]

}
\keyword{internal}
